the starting point of Maxq and Maxl has x_i = - i - 1 computed in double, the unsigned index making the last ten components about 1.8e19

+++ src/LukFiBlock.cpp
@@ -823,14 +823,14 @@ void LukFiBlock::SetInitialPoint( void )
    for( decltype( x.size() ) i = 0 ; i < 10 ; i++ )
     x[ i ].set_value( i + 1 );
    for( decltype( x.size() ) i = 10 ; i < 20 ; i++ )
-    x[ i ].set_value( -i - 1 );
+    x[ i ].set_value( - double( i ) - 1 );
    break;
   // Maxl    - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   case( 15 ):
    for( decltype( x.size() ) i = 0 ; i < 10 ; i++ )
     x[ i ].set_value( i + 1 );
    for( decltype( x.size() ) i = 10 ; i < 20 ; i++ )
-    x[ i ].set_value( -i - 1 );
+    x[ i ].set_value( - double( i ) - 1 );
    break;
   // TR48    - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   case( 16 ):
